[FIX] stock: quant domain in valorisation
- Create a Customer location (CL) and set a Company on it
- Configure Product A in FIFO / Perpetual
- Receive 2 Unit(s) of A in Stock
- Create an Internal Transfer
  1 Unit(s) of Product A
  Source location: Stock
  Destination Location: CL
- Validate

The Inventory Valuation report incorrectly retrieves a quantity of 2 in
stock instead of 1.

The error is first due to a misconfiguration: the CL should not have a
Company set. However, we should not take into account any quantity in a
customer location, even if there is a Company set.

opw-1884675
opw-1884835

closes odoo/odoo#27145