[FIX] mass_mailing: make campaign default to non-A/B testing
In A/B testing mode, a given recipient will only ever receive a single
email from a given mass-mailing campaign, no matter how many mailings
were sent to them.

This is useful for A/B-testing various mailings in order to check their
results. But is very difficult to understand for users who simply want
to organize their mailings into campaigns.

Additionally, the A/B testing option is hidden as a technical feature, so
it is hard to discover in order to troubleshoot unexpected number of
emails sent by a given mailing.

It will be safer to keep it off by default.
(keeping the parameter explicit, to better show that it is on purpose)

This partially reverts eb6e141dde793ab2442c46851eabb4c150d8cf68