[FIX] stock: package to master data
commit a7525b1db5affd907b3c17d6632a3746a8c75237 moved package menu to
operation and rename it delivery package. However move it to opeation
could confuse the user and let him imagine that he is able to create the
package content from this menu which is not the case. This menu only
allow you to see the package, its location and its content.

Also package are not always delivery package since it could be use in
internal picking type so it's better to remove the 'delivery' from the
menu item.