[IMP] hw_scale: fix to avoid conflicts with other driver
There are two scales that are supported, but the Adam scale
possibly won't return any answer so there is no way to detect then
if it is an Adam scale or something else.  As we have not implemented
a detection method for the caliper driver yet, we exclude it from
the devices the scale can detect.  (by comparing /dev/serial/by-id
and by-path)