[FIX] web: 'Shortcuts' in user menu (community)
Revision on https://github.com/odoo/odoo/commit/d5fb61f1cd246a319a5c5632c4592b8428e864d5

This menu item should not appear in the community version of Odoo.
The commit above mistakenly added this menu item, even though no
logic was implemented for it. This was a consequence of the copy
& paste approach to rapidly deploy the new navbar in community.

(Note: a big refactoring of the web client is coming very soon,
so this duplicated code issue should only be temporarily).

This commit fixes the issue by removing this menu item in the
community version.