[FIX] delivery: new error message
Making a delivery with an external carrier without setting any partner
will sometimes lead to a crash on the API respond from the carrier. This technical error is sent back to the user without any proper sentence human readable.

This commit set the field 'partner_id' required in case of delivery with external carrier to avoid any technical error

Task : 1886733