[IMP] fleet: add activity view on main actions
Purpose is to extend the use of the newly-introduced activity view allowing
to see at a glance activities to perform on a given model. It eases daily
job of people working with activities.

Updated :

  * fleet vehicle and contract main menu actions;

Some menu entries are also renamed to ensure coherence in all menu entries
that all use vehicles, with an 's'.

This commit is linked to task ID 1889413.