[FIX] web_settings_dashboard: set up your company
Before this commit, it was not possible to set up a company.
This is caused by the clickable DOM element being a `<a>` with
`href='#'`. This redirects to the 1st app in community (usually
the 'Discuss' app), which prevents opening the form view to
set up information about your company.

This commit fixes the issue by preventing the browser to change
the URL in the address bar when clicking on 'Set up'.

Closes #27434
Closes #27442